+++ includes/reader-revenue/my-account/class-woocommerce-my-account.php
@@ -37,6 +37,11 @@ class WooCommerce_My_Account {
 	];
 
 	/**
+	 * Cron hook for syncing email change with ESP.
+	 */
+	const SYNC_ESP_EMAIL_CHANGE_CRON_HOOK = 'newspack_esp_sync_email_change';
+
+	/**
 	 * Initialize.
 	 *
 	 * @codeCoverageIgnore
@@ -73,6 +78,7 @@ class WooCommerce_My_Account {
 			\add_filter( 'wc_memberships_members_area_my-memberships_actions', [ __CLASS__, 'hide_cancel_button_from_memberships_table' ] );
 			\add_filter( 'wc_memberships_my_memberships_column_names', [ __CLASS__, 'remove_next_bill_on' ], 21 );
 			\add_action( 'profile_update', [ __CLASS__, 'handle_admin_email_change_request' ], 10, 3 );
+			\add_action( self::SYNC_ESP_EMAIL_CHANGE_CRON_HOOK, [ __CLASS__, 'sync_email_change_with_esp' ], 10, 3 );
 		}
 	}
 
@@ -853,43 +859,50 @@ class WooCommerce_My_Account {
 				\wc_add_notice( __( 'Something went wrong. Please try again.', 'newspack-plugin' ), 'error' );
 			} else {
 				$sent = [];
-				foreach ( [ $old_email, $new_email ] as $email ) {
-					if (
-						Emails::send_email(
-							Reader_Activation_Emails::EMAIL_TYPES['CHANGE_EMAIL'],
-							$email,
+				if (
+					Emails::send_email(
+						Reader_Activation_Emails::EMAIL_TYPES['CHANGE_EMAIL_CANCEL'],
+						$old_email,
+						[
 							[
-								[
-									'template' => '*EMAIL_VERIFICATION_URL*',
-									'value'    => self::get_email_change_url( self::VERIFY_EMAIL_CHANGE_PARAM, $old_email ),
-								],
-								[
-									'template' => '*EMAIL_CANCELLATION_URL*',
-									'value'    => self::get_email_change_url( self::CANCEL_EMAIL_CHANGE_PARAM, $old_email ),
-								],
-							]
-						)
-					) {
-						$sent[] = $email;
-					}
+								'template' => '*PENDING_EMAIL_ADDRESS*',
+								'value'    => $new_email,
+							],
+							[
+								'template' => '*EMAIL_CANCELLATION_URL*',
+								'value'    => self::get_email_change_url( self::CANCEL_EMAIL_CHANGE_PARAM, $old_email ),
+							],
+						]
+					)
+				) {
+					$sent[] = $old_email;
+				}
+				if (
+					Emails::send_email(
+						Reader_Activation_Emails::EMAIL_TYPES['CHANGE_EMAIL'],
+						$new_email,
+						[
+							[
+								'template' => '*EMAIL_VERIFICATION_URL*',
+								'value'    => self::get_email_change_url( self::VERIFY_EMAIL_CHANGE_PARAM, $old_email ),
+							],
+							[
+								'template' => '*EMAIL_CANCELLATION_URL*',
+								'value'    => self::get_email_change_url( self::CANCEL_EMAIL_CHANGE_PARAM, $old_email ),
+							],
+						]
+					)
+				) {
+					$sent[] = $new_email;
 				}
 				if ( empty( $sent ) ) {
 					\wc_add_notice( __( 'Something went wrong. Please contact the site administrator.', 'newspack-plugin' ), 'error' );
-				} elseif ( count( $sent ) === 1 ) {
+				} else {
 					\wc_add_notice(
 						sprintf(
 							// Translators: %s is the email address the verification email was sent to..
 							__( 'A verification email has been sent to %s. Please verify to complete the change.', 'newspack-plugin' ),
-							$sent[0]
-						)
-					);
-				} else {
-					\wc_add_notice(
-						sprintf(
-							// Translators: 1 and 2 are the email addresses the verification email was sent to.
-							__( 'A verification email has been sent to %1$s and %2$s. Please verify to complete the change.', 'newspack-plugin' ),
-							$sent[0],
-							$sent[1]
+							$new_email
 						)
 					);
 				}
@@ -915,7 +928,7 @@ class WooCommerce_My_Account {
 		$old_email = $user->user_email;
 		if ( $new_email !== $old_email && \is_email( $new_email ) && \is_email( $old_email ) ) {
 			self::maybe_sync_email_change_with_stripe( $user_id, $new_email );
-			ESP_Sync::sync_email_change( $user_id, $new_email, $old_email );
+			self::sync_email_change_with_esp( $user_id, $new_email, $old_email );
 		}
 	}
 
@@ -930,7 +943,8 @@ class WooCommerce_My_Account {
 		if ( ! $secret ) {
 			return;
 		}
-		$error     = __( 'Something went wrong.', 'newspack-plugin' );
+		$message   = __( 'Your email address has been successfully updated.', 'newspack-plugin' );
+		$is_error  = false;
 		$user_id   = \get_current_user_id();
 		$new_email = \get_user_meta( $user_id, self::PENDING_EMAIL_CHANGE_META, true );
 		$old_email = \wp_get_current_user()->user_email;
@@ -947,16 +961,29 @@ class WooCommerce_My_Account {
 				$customer->set_billing_email( $new_email );
 				$customer->save();
 				self::maybe_sync_email_change_with_stripe( $user_id, $new_email );
-				ESP_Sync::sync_email_change( $user_id, $new_email, $old_email );
+				self::sync_email_change_with_esp( $user_id, $new_email, $old_email );
 				\delete_user_meta( $user_id, self::PENDING_EMAIL_CHANGE_META );
-				\wc_add_notice( __( 'Your email address has been successfully updated.', 'newspack-plugin' ) );
 			} else {
-				\wc_add_notice( $error, 'error' );
+				$message  = __( 'Something went wrong.', 'newspack-plugin' );
+				$is_error = true;
 			}
 		} else {
-			\wc_add_notice( $error, 'error' );
+			$message  = __( 'This email change request has been cancelled or expired.', 'newspack-plugin' );
+			$is_error = true;
 		}
-		\wp_safe_redirect( \wc_get_endpoint_url( 'edit-account', '', \wc_get_page_permalink( 'myaccount' ) ) );
+		\wp_safe_redirect(
+			\add_query_arg(
+				[
+					'message'  => $message,
+					'is_error' => $is_error,
+				],
+				\wc_get_endpoint_url(
+					'edit-account',
+					'',
+					\wc_get_page_permalink( 'myaccount' )
+				)
+			)
+		);
 		exit;
 	}
 
@@ -972,13 +999,27 @@ class WooCommerce_My_Account {
 			return;
 		}
 		$current_email = \wp_get_current_user()->user_email;
+		$message       = __( 'Your email address change request has been cancelled.', 'newspack-plugin' );
+		$is_error      = false;
 		if ( \wp_hash( $current_email ) === $secret ) {
 			\delete_user_meta( \get_current_user_id(), self::PENDING_EMAIL_CHANGE_META );
-			\wc_add_notice( __( 'Your email change request has been cancelled.', 'newspack-plugin' ) );
 		} else {
-			\wc_add_notice( __( 'Something went wrong.', 'newspack-plugin' ), 'error' );
+			$message  = __( 'This email change request has been cancelled or expired.', 'newspack-plugin' );
+			$is_error = true;
 		}
-		\wp_safe_redirect( \wc_get_endpoint_url( 'edit-account', '', \wc_get_page_permalink( 'myaccount' ) ) );
+		\wp_safe_redirect(
+			\add_query_arg(
+				[
+					'message'  => $message,
+					'is_error' => $is_error,
+				],
+				\wc_get_endpoint_url(
+					'edit-account',
+					'',
+					\wc_get_page_permalink( 'myaccount' )
+				)
+			)
+		);
 		exit;
 	}
 
@@ -989,14 +1030,40 @@ class WooCommerce_My_Account {
 	 * @param string $email   New email.
 	 */
 	public static function maybe_sync_email_change_with_stripe( $user_id, $email ) {
-		$request = Stripe_Connection::update_customer_data(
+		$result = Stripe_Connection::update_customer_data(
 			$user_id,
 			[
 				'email' => $email,
 			]
 		);
-		if ( \is_wp_error( $request ) ) {
+		if ( false === $result ) {
+			Logger::log( 'Skipping Stripe email update: no Stripe customer found for user ' . $email );
+		}
+		if ( \is_wp_error( $result ) ) {
 			Logger::error( 'Error updating Stripe customer email: ' . $result->get_error_message() );
+		}
+	}
+
+	/**
+	 * Sync email change with site ESPs.
+	 *
+	 * @param int    $user_id User ID.
+	 * @param string $new_email New email address.
+	 * @param string $old_email Old email address.
+	 */
+	public static function sync_email_change_with_esp( $user_id, $new_email, $old_email ) {
+		if ( ! ESP_Sync::can_esp_sync() ) {
+			return;
+		}
+		$contact = ESP_Sync::get_contact_data( $user_id );
+		if ( ! $contact ) {
+			return;
+		}
+		$update = ESP_Sync::sync( $contact, 'Email_Change', array_merge( $contact, [ 'email' => $old_email ] ) );
+		if ( is_wp_error( $update ) ) {
+			// If the update failed, retry in 24 hours.
+			\wp_schedule_single_event( time() + DAY_IN_SECONDS, self::SYNC_ESP_EMAIL_CHANGE_CRON_HOOK, [ $user_id, $new_email, $old_email ] );
+			Logger::error( 'Error syncing email change with ESP: ' . $update->get_error_message() . '. Retrying in 24 hours.' );
 		}
 	}