+++ src/wizards/audience/views/campaigns/index.js
@@ -8,7 +8,7 @@ import '../../../../shared/js/public-path';
 /**
  * WordPress dependencies.
  */
-import { Component  } from '@wordpress/element';
+import { Component } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { addQueryArgs } from '@wordpress/url';